% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_function2.R
\name{tmpfct2}
\alias{tmpfct2}
\title{One Line Title}
\usage{
tmpfct2(x, y)
}
\arguments{
\item{x}{x.}

\item{y}{y.}
}
\value{
Specify what are returned.
}
\description{
One paragraph description
}
\details{
Details
(Include subjects for verbs.)
(Use 3rd person forms for verbs.)
}
\examples{
tmpfct2(1, 2)

}
